+++ src/paimon/common/memory/memory_segment_utils.h
@@ -17,6 +17,7 @@
 #pragma once
 
 #include <algorithm>
+#include <cstddef>
 #include <cstdint>
 #include <cstring>
 #include <memory>
@@ -41,7 +42,7 @@ class PAIMON_EXPORT MemorySegmentUtils {
     ~MemorySegmentUtils() = delete;
 
     /// Allocate bytes in pool
-    static std::shared_ptr<Bytes> AllocateBytes(int32_t length, MemoryPool* pool);
+    static std::shared_ptr<Bytes> AllocateBytes(size_t length, MemoryPool* pool);
 
     /// Copy target segments from source byte[].
     ///
